+++ sql/cache_test.go
@@ -15,11 +15,10 @@
 package sql
 
 import (
-	"context"
 	"errors"
 	"testing"
 
-	"github.com/stretchr/testify/require"
+		"github.com/stretchr/testify/require"
 )
 
 func TestLRUCache(t *testing.T) {
@@ -176,37 +175,5 @@ func TestRowsCache(t *testing.T) {
 		require.NoError(cache.Add(Row{1, "foo"}))
 		require.Len(cache.Get(), 1)
 		require.True(freed)
-	})
-}
-
-func BenchmarkHashOf(b *testing.B) {
-	ctx := context.Background()
-	row := NewRow(1, "1")
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
-		sum, err := HashOf(ctx, row)
-		if err != nil {
-			b.Fatal(err)
-		}
-		if sum != 11268758894040352165 {
-			b.Fatalf("got %v", sum)
-		}
-	}
-}
-
-func BenchmarkParallelHashOf(b *testing.B) {
-	ctx := context.Background()
-	row := NewRow(1, "1")
-	b.ResetTimer()
-	b.RunParallel(func(pb *testing.PB) {
-		for pb.Next() {
-			sum, err := HashOf(ctx, row)
-			if err != nil {
-				b.Fatal(err)
-			}
-			if sum != 11268758894040352165 {
-				b.Fatalf("got %v", sum)
-			}
-		}
 	})
 }
